// src/features/navigation/components/overlay/types.ts
export type Loose = any;

export interface MapUIOverlayProps {
  backendOnline: boolean;
  navigating: boolean;
  driveSegments: Loose;
  searchTop: number;
  customOriginName: string;
  handleDestinationSelect: Loose;
  handleClear: Loose;
  handleOriginChange: Loose;
  tunnelWarning: string | null;
  insets: Loose;
  handleTunnelWarningDismiss: Loose;
  stepToShow: Loose;
  nextStep: Loose;
  distToTurn: number | null;
  currentLanes: Loose[];
  aheadEvents: Loose[];
  truckSituation: Loose;
  displayLanes: Loose[];
  laneGlowBg: Loose;
  laneGlowShadow: Loose;
  route: Loose;
  routeControlsVisible: boolean;
  setRouteControlsVisible: Loose;
  navPhase: Loose;
  routeAheadPOIs: Loose[];
  handleRouteTimelinePOIPress: Loose;
  trafficSegments: Loose[];
  fasterOffer: Loose;
  handleAcceptFasterRoute: Loose;
  dismissOffer: Loose;
  optionsOpen: boolean;
  setOptionsOpen: Loose;
  mapMode: Loose;
  setMapMode: Loose;
  lightMode: boolean;
  setLightMode: Loose;
  voiceMuted: boolean;
  setVoiceMuted: Loose;
  mapLayers: Loose;
  toggleLayer: Loose;
  avoidUnpaved: boolean;
  setAvoidUnpaved: Loose;
  simulating: boolean;
  startSim: Loose;
  stopSim: Loose;
  poiCategory: Loose;
  handlePOISearch: Loose;
  sarMode: boolean;
  handleSARSearch: Loose;
  googleUser: Loose;
  setShowAccountModal: Loose;
  starredPOIs: Loose[];
  setBorderCrossings: Loose;
  setShowBorderPanel: Loose;
  isSearchingAlongRoute: boolean;
  handleSearchAlongRoute: Loose;
  setMapIsLoaded: Loose;
  userCoords: [number, number] | null;
  handleReportCamera: Loose;
  gpsReady: boolean;
  rerouting: boolean;
  loadingRoute: boolean;
  showBorderPanel: boolean;
  borderCrossings: Loose[];
  profile: Loose;
  poiResults: Loose[];
  loadingPOI: boolean;
  clearPOI: Loose;
  handlePOINavigate: Loose;
  parkingResults: Loose[];
  parkingSource?: 'gpt' | 'route' | null;
  setParkingResults: Loose;
  setParkingSource: Loose;
  remainingDriveMin?: number;
  speedKmh?: number;
  urgentParkingResults: Loose[];
  navigateTo: Loose;
  addWaypoint: Loose;
  setSelectedParking: Loose;
  navigation: Loose;
  fuelResults: Loose[];
  setFuelResults: Loose;
  tachographResult: Loose;
  tachoSummary: Loose;
  bluetoothTacho: BluetoothTachoState;
  setTachographResult: Loose;
  businessResults: Loose[];
  setBusinessResults: Loose;
  routeOptions: Loose[];
  selectedRouteIdx: number | null;
  routeOptDest: Loose;
  restrictionChecking: boolean;
  restrictionWarnings: string[];
  handleSelectRouteOption: Loose;
  setRouteOptions: Loose;
  setRouteOptDest: Loose;
  setSelectedRouteIdx: Loose;
  setRestrictionWarnings: Loose;
  handleStartRoute: Loose;
  drivingSeconds: number;
  handleExportGPX: Loose;
  selectedParking: Loose;
  selectedFuel: Loose;
  setSelectedFuel: Loose;
  wakeWordHeard: boolean;
  currentStep: number;
  speed: number;
  speedLimit: number | null;
  remainingSeconds: number;
  destination: [number, number] | null;
  destinationName: string;
  handleStopNav: Loose;
  handleStart: Loose;
  buildElevProfile: Loose;
  fetchWeatherForRoute: Loose;
  destinationRef: MutableRefObject<[number, number] | null>;
  destinationNameRef: MutableRefObject<string>;
  waypointsRef: MutableRefObject<[number, number][]>;
  dominantCongestion: Loose;
  elevProfile: number[];
  weatherPoints: Loose[];
  departLabel: Loose;
  pickDeparture: Loose;
  waypoints: [number, number][];
  waypointNames: string[];
  setWaypoints: Loose;
  setWaypointNames: Loose;
  speedingBg: Loose;
  proximityAlerts: Loose;
  nearestParkingM: number | null;
  hillWarnings: Loose[];
  cameraAlert: Loose;
  cameraFlashAnim: Loose;
  mapPitch: number;
  setMapPitch: Loose;
  cameraRef: MutableRefObject<any>;
  geminiChatOpen: boolean;
  gptChatOpen: boolean;
  setGeminiChatOpen: Loose;
  setGptChatOpen: Loose;
  isTracking: boolean;
  setIsTracking: Loose;
  suppressPanUntilRef: MutableRefObject<number>;
  gptHistory: Loose[];
  geminiHistory: Loose[];
  chatInput: string;
  setChatInput: Loose;
  gptLoading: boolean;
  geminiLoading: boolean;
  handleChat: Loose;
  handleTargetedQuickAction: Loose;
  isRecording: boolean;
  handleMicStart: Loose;
  handleMicStop: Loose;
  kbHeight: number;
  gptScrollRef: Loose;
  geminiScrollRef: Loose;
  micLoading: boolean;
  showAccountModal: boolean;
  setGoogleUser: Loose;
  isMountedRef: MutableRefObject<boolean>;
  setStarredPOIs: Loose;
}